Switch to MySQL, add in authentication script and manual setup utility file
Note: shouldn't need manual setup feature with a couple modifications to the database create_db() function. TODO: Add in higher privileged credentials for initial setup as arguments for database initialization if I decide to encapsulate this feature. Investigate if this is a preferred method. Also, testing script will no longer work for this push. Modifications needed with respect to DB object initialization.